requirements.txt
astroid==2.3.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cssselect==1.1.0
dnspython==1.16.0
fake-useragent==0.1.11
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.19
Pillow==7.0.0
protobuf==3.6.1
pylint==2.4.4
pyquery==1.4.1
python-dotenv==0.10.5
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==0.16.1
wrapt==1.11.2
xlwt==1.3.0
